refactor: Update customer.controller and customer.service.interface
This commit refactors the `customer.controller.ts` and `customer.service.interface.ts` files.

In the `customer.controller.ts` file, the following changes were made:
- Added new API endpoints for creating, updating, and deleting customers.
- Updated the `getCustomer` endpoint to return a `CustomerResponseDto` instead of a `CustomerDto`.

In the `customer.service.interface.ts` file, the following changes were made:
- Updated the `findById` method to return a `CustomerResponseDto` instead of a `CustomerDto`.
- Updated the `create` method to accept a `CreateCustomerDto` instead of a `CustomerDto`.
- Updated the `update` method to accept an `UpdateCustomerDto` instead of a `CustomerDto`.

These changes improve the functionality and maintainability of the customer-related operations in the application.